pyproject.toml
[project]
name = "sorabot"
version = "1.0.0"
description = "基于 NoneBot2 开发，互通多平台，超可爱的林汐酱"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0"}
authors = [{name = "Komorebi", email = "mute231010@gmail.com"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "nonebot2[fastapi]>=2.0.1",
    "tortoise>=0.1.1",
    "tortoise-orm>=0.20.0",
    "ujson>=5.8.0",
    "ruamel-yaml>=0.17.32",
    "apscheduler>=3.10.3",
    "httpx>=0.24.1",
    "aiofiles>=23.2.1",
    "retrying>=1.3.4",
    "rich>=13.5.2",
    "GitPython>=3.1.32",
    "nonebot-plugin-send-anything-anywhere>=0.3.0",
    "nonebot-plugin-alconna>=0.16.0",
    "nonebot-plugin-userinfo>=0.1.3",
    "pillow>=10.0.0",
    "websockets>=11.0.3",
    "pandas>=2.0.3",
    "toml>=0.10.2",
    "dnspython>=2.5.0",
]

[project.urls]
homepage = "https://bot.netsora.info/"
repository = "https://github.com/netsora/SoraBot"
documentation = "https://bot.netsora.info/"

[project.optional-dependencies]
adapters = [
    "nonebot-adapter-onebot>=2.3.1",
    "nonebot-adapter-qq>=1.3.5",
    "nonebot-adapter-telegram>=0.1.0b15",
]

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "ruff>=0.0.284",
    "pre-commit>=3.3.3",
    "isort>=5.13.2",
]

[tool.pdm.scripts]
start = "pdm run bot.py"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]


[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]
line-length = 120
target-version = "py310"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
executionEnvironments = [
    { root = "./tests", extraPaths = ["./"] },
    { root = "./" },
]
typeCheckingMode = "basic"
reportShadowedImports = false

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
